chore: Removed deprecated UniqueEntity constraint fron roadiz

+++ lib/Rozier/src/Controllers/NodeTypes/NodeTypesUtilsController.php
@@ -83,6 +83,7 @@ class NodeTypesUtilsController extends RozierApp
      * @param Request $request
      *
      * @return BinaryFileResponse
+     * @throws RuntimeError
      */
     public function exportDocumentationAction(Request $request): BinaryFileResponse
     {
@@ -91,6 +92,10 @@ class NodeTypesUtilsController extends RozierApp
         $documentationGenerator = new DocumentationGenerator($this->nodeTypesBag, $this->getTranslator());
 
         $tmpfname = tempnam(sys_get_temp_dir(), date('Y-m-d-H-i-s') . '.zip');
+        if (false === $tmpfname) {
+            throw new RuntimeError('Unable to create temporary file.');
+        }
+
         unlink($tmpfname); // Deprecated: ZipArchive::open(): Using empty file as ZipArchive is deprecated
         $zipArchive = new ZipArchive();
         $zipArchive->open($tmpfname, ZipArchive::CREATE);
@@ -161,6 +166,9 @@ class NodeTypesUtilsController extends RozierApp
 
         $zipArchive = new ZipArchive();
         $tmpfname = tempnam(sys_get_temp_dir(), date('Y-m-d-H-i-s') . '.zip');
+        if (false === $tmpfname) {
+            throw new RuntimeError('Unable to create temporary file.');
+        }
         unlink($tmpfname); // Deprecated: ZipArchive::open(): Using empty file as ZipArchive is deprecated
         $zipArchive->open($tmpfname, ZipArchive::CREATE);
 
@@ -212,6 +220,9 @@ class NodeTypesUtilsController extends RozierApp
 
             if ($file->isValid()) {
                 $serializedData = file_get_contents($file->getPathname());
+                if (false === $serializedData) {
+                    throw new RuntimeError('Unable to read uploaded file.');
+                }
 
                 if (null !== json_decode($serializedData)) {
                     $this->nodeTypesImporter->import($serializedData);
